moved contants out of context and abstracted layout into another component
