+++ src/index.js
@@ -15,7 +15,6 @@ import thunk from 'redux-thunk'
 import rootReducer from 'redux/reducers.js'
 
 import { ConnectionStringProvider } from 'context/ConnectionStringContext'
-import { EventsPageCtxProvider } from 'context/EventsInfo'
 import App from 'App.jsx'
 import logger from './utils/logger'
 
@@ -63,9 +62,7 @@ function render() {
       <Provider store={store}>
         <MuiThemeProvider theme={theme}>
           <ConnectionStringProvider>
-            <EventsPageCtxProvider>
-              <App history={history} />
-            </EventsPageCtxProvider>
+            <App history={history} />
           </ConnectionStringProvider>
         </MuiThemeProvider>
       </Provider>
